Make commitments service employer friendly be removing provider id from get draft apprentice details method

+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Models/DraftApprenticeshipViewModel.cs
@@ -22,7 +22,7 @@ namespace SFA.DAS.ProviderCommitments.Web.Models
             EndDate = new MonthYearModel("");
         }
 
-        public int ProviderId { get; set; }
+        public long ProviderId { get; set; }
         public string CohortReference { get; set; }
         public long? CohortId { get; set; }
